modules/face_enroller/requirements.txt
opencv-python-headless
insightface
torch
torchvision
numpy
pymilvus
python-dotenv
onnx
onnxruntime
